Replace 'game-install' signal with a straight method call.
Still using the default application object here.
