+++ lutris/game.py
@@ -61,7 +61,6 @@ class Game(GObject.Object):
         "game-stopped": (GObject.SIGNAL_RUN_FIRST, None, ()),
         "game-removed": (GObject.SIGNAL_RUN_FIRST, None, ()),
         "game-updated": (GObject.SIGNAL_RUN_FIRST, None, ()),
-        "game-install": (GObject.SIGNAL_RUN_FIRST, None, ()),
         "game-install-update": (GObject.SIGNAL_RUN_FIRST, None, ()),
         "game-install-dlc": (GObject.SIGNAL_RUN_FIRST, None, ()),
         "game-installed": (GObject.SIGNAL_RUN_FIRST, None, ()),
@@ -333,6 +332,40 @@ class Game(GObject.Object):
             if not self.compositor_disabled:
                 disable_compositing()
                 self.compositor_disabled = True
+
+    def install(self, launch_ui_delegate=None):
+        application = Gio.Application.get_default()
+        if not launch_ui_delegate:
+            launch_ui_delegate = application.launch_ui_delegate
+
+        if not self.slug:
+            raise ValueError("Invalid game passed: %s" % self)
+
+        """Request installation of a game"""
+        if not self.service or self.service == "lutris":
+            application.show_lutris_installer_window(game_slug=self.slug)
+            return
+
+        service = launch_ui_delegate.get_service(self.service)
+        db_game = service.get_service_db_game(self)
+        if not db_game:
+            logger.error("Can't find %s for %s", self.name, service.name)
+            return
+
+        try:
+            game_id = service.install(db_game)
+        except ValueError as e:
+            logger.debug(e)
+            game_id = None
+
+        if game_id:
+            def on_error(_game, error):
+                logger.exception("Unable to install game: %s", error)
+                return True
+
+            game = Game(game_id)
+            game.connect("game-error", on_error)
+            game.launch(launch_ui_delegate)
 
     def remove(self, delete_files: bool = False, no_signal: bool = False) -> None:
         """Uninstall a game, and removes it from the library if this it has no playtime.
